frigate_identity_service/requirements-cpu.txt
paho-mqtt==2.1.0
requests>=2.31.0
# torch and torchvision are installed separately in the Dockerfile
# using --index-url https://download.pytorch.org/whl/cpu for CPU-only builds
torchreid>=0.2.5
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0
sqlalchemy>=2.0.0
